Move files to proper packages
