+++ packages/apps/job-launcher/server/src/modules/payment/providers/payment-provider.abstract.ts
@@ -1,58 +1,13 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { VatType } from '../../common/enums/payment';
-
-export interface PaymentMethod {
-  id: string;
-  brand: string;
-  last4: string;
-  expMonth: number;
-  expYear: number;
-  default: boolean;
-}
-
-export interface BillingAddress {
-  line1?: string;
-  city?: string;
-  country?: string;
-  postal_code?: string;
-}
-
-export interface CustomerData {
-  email: string;
-  name?: string;
-  address?: BillingAddress;
-  default_payment_method?: string;
-}
-
-export interface TaxId {
-  id: string;
-  type: VatType;
-  value: string;
-}
-
-export interface Invoice {
-  id: string;
-  payment_intent: string | null;
-  status?: string;
-  amount_due: number;
-  currency: string;
-}
-
-export interface SetupIntent {
-  customer: string;
-  payment_method: string;
-}
-
-export interface PaymentIntent {
-  customer: string;
-  id: string;
-  client_secret: string;
-  status: string;
-  amount: number;
-  amount_received: number;
-  currency: string;
-  latest_charge: string;
-}
+import { VatType } from '../../../common/enums/payment';
+import {
+  Invoice,
+  PaymentIntent,
+  CustomerData,
+  PaymentMethod,
+  TaxId,
+  SetupIntent,
+} from '../payment.interface';
 
 @Injectable()
 export abstract class PaymentProvider {
