+++ packages/apps/job-launcher/server/src/modules/payment/providers/stripe/stripe.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { StripeService } from './stripe.service';
-import { StripeConfigService } from '../../common/config/stripe-config.service';
+import { StripeConfigService } from '../../../../common/config/stripe-config.service';
 
 @Module({
   providers: [StripeService, StripeConfigService],
